move executor commitment state to executor module
